+++ src/main/java/mods/belgabor/amtweaker/mods/amt/AMT.java
@@ -1,15 +1,13 @@
 package mods.belgabor.amtweaker.mods.amt;
 
 import minetweaker.MineTweakerAPI;
-import mods.belgabor.amtweaker.mods.amt.handlers.Evaporator;
-import mods.belgabor.amtweaker.mods.amt.handlers.IceMaker;
-import mods.belgabor.amtweaker.mods.amt.handlers.Plate;
-import mods.belgabor.amtweaker.mods.amt.handlers.Processor;
+import mods.belgabor.amtweaker.mods.amt.handlers.*;
 
 public class AMT {
     public AMT() {
         MineTweakerAPI.registerClass(Evaporator.class);
         MineTweakerAPI.registerClass(Plate.class);
+        MineTweakerAPI.registerClass(Pan.class);
         MineTweakerAPI.registerClass(Processor.class);
         MineTweakerAPI.registerClass(IceMaker.class);
     }