+++ changelog.md
@@ -5,6 +5,7 @@ Version 0.1alpha
 ----------------
 Early alpha version
 
+* Clay Pan (add/remove recipe, add heat source)
 * Cooking Iron Plate (add/remove recipe, add heat source)
 * Evaporator (add/remove recipe)
 * Food Processor/Jaw Crusher (add/remove recipe)